+++ src/Funds/constants.js
@@ -15,10 +15,11 @@ export const FUND_STATUSES = {
   FROZEN: 'Frozen',
 };
 
+// use already formatted labels in selections
 export const FUND_STATUSES_OPTIONS = [
-  { label: <FormattedMessage id="ui-finance.fund.status.active" />, value: FUND_STATUSES.ACTIVE },
-  { label: <FormattedMessage id="ui-finance.fund.status.inactive" />, value: FUND_STATUSES.INACTIVE },
-  { label: <FormattedMessage id="ui-finance.fund.status.frozen" />, value: FUND_STATUSES.FROZEN },
+  { labelId: 'ui-finance.fund.status.active', value: FUND_STATUSES.ACTIVE },
+  { labelId: 'ui-finance.fund.status.inactive', value: FUND_STATUSES.INACTIVE },
+  { labelId: 'ui-finance.fund.status.frozen', value: FUND_STATUSES.FROZEN },
 ];
 
 export const FUND_FILTERS = {