+++ src/Groups/constants.js
@@ -21,14 +21,14 @@ export const GROUP_STATUS = {
   inactive: 'Inactive',
 };
 
-export const GROUP_STATUS_LABEL = {
-  [GROUP_STATUS.active]: <FormattedMessage id="ui-finance.groups.status.active" />,
-  [GROUP_STATUS.frozen]: <FormattedMessage id="ui-finance.groups.status.frozen" />,
-  [GROUP_STATUS.inactive]: <FormattedMessage id="ui-finance.groups.status.inactive" />,
+export const GROUP_STATUS_LABEL_ID = {
+  [GROUP_STATUS.active]: 'ui-finance.groups.status.active',
+  [GROUP_STATUS.frozen]: 'ui-finance.groups.status.frozen',
+  [GROUP_STATUS.inactive]: 'ui-finance.groups.status.inactive',
 };
 
 export const GROUP_STATUS_OPTIONS = Object.values(GROUP_STATUS).map(status => ({
-  label: GROUP_STATUS_LABEL[status],
+  labelId: GROUP_STATUS_LABEL_ID[status],
   value: status,
 }));